# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI CLI Syncer API
  description: |
    API for managing AI CLI tools (Claude, Cursor, Codex, Gemini, etc.) MCP configurations and Rules.
    This API enables synchronization of Model Context Protocol (MCP) servers and coding rules across multiple AI development tools.
  version: 0.1.0
  contact:
    name: AI CLI Syncer
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
...
paths:
...
  /mcp/master:
    get:
      tags: [MCP]
      summary: Get master MCP configuration
      operationId: getMasterMcp
      responses:
        '200':
          description: Master MCP config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterMcpConfig'
    post:
      tags: [MCP]
      summary: Save master MCP configuration
      operationId: saveMasterMcp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MasterMcpConfig'
      responses:
        '200':
          description: MCP config saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /mcp/definitions:
    get:
      tags: [MCP Definitions]
      summary: Get MCP definitions
      operationId: getMcpDefinitions
      responses:
        '200':
          description: List of MCP definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpDefinition'
    post:
      tags: [MCP Definitions]
      summary: Create MCP definition
      operationId: createMcpDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMcpDefinitionRequest'
      responses:
        '201':
          description: Definition created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpDefinition'
  /mcp/definitions/{id}:
    put:
      tags: [MCP Definitions]
      summary: Update MCP definition
      operationId: updateMcpDefinition
      parameters:
        - $ref: '#/components/parameters/McpId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMcpDefinitionRequest'
      responses:
        '200':
          description: Definition updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpDefinition'
    delete:
      tags: [MCP Definitions]
      summary: Delete MCP definition
      operationId: deleteMcpDefinition
      parameters:
        - $ref: '#/components/parameters/McpId'
      responses:
        '200':
          description: Definition deleted
  /mcp/sets:
    get:
      tags: [MCP Sets]
      summary: Get MCP sets
      operationId: getMcpSets
      responses:
        '200':
          description: List of MCP sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpSet'
    post:
      tags: [MCP Sets]
      summary: Create MCP set
      operationId: createMcpSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMcpSetRequest'
      responses:
        '201':
          description: Set created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpSet'
  /mcp/sets/{id}:
    put:
      tags: [MCP Sets]
      summary: Update MCP set
      operationId: updateMcpSet
      parameters:
        - $ref: '#/components/parameters/McpSetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMcpSetRequest'
      responses:
        '200':
          description: Set updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpSet'
    delete:
      tags: [MCP Sets]
      summary: Delete MCP set
      operationId: deleteMcpSet
      parameters:
        - $ref: '#/components/parameters/McpSetId'
      responses:
        '200':
          description: Set deleted
...
  /config:
    get:
      tags: [Config]
      summary: Get config file content
      operationId: getConfig
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Path to config file
      responses:
        '200':
          description: Config content
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Config]
      summary: Save config file
      operationId: saveConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveConfigRequest'
      responses:
        '200':
          description: Config saved
...
components:
  parameters:
...
    McpId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: MCP definition UUID
    McpSetId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: MCP set UUID
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    McpServer:
      type: object
      properties:
        command:
          type: string
          minLength: 1
        args:
          type: array
          items:
            type: string
        description:
          type: string
        category:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
      required: [command, args]
    MasterMcpConfig:
      type: object
      properties:
        mcpServers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/McpServer'
      required: [mcpServers]
    McpDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        description:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
      required: [id, name, command, args]
    CreateMcpDefinitionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        command:
          type: string
          minLength: 1
        args:
          type: array
          items:
            type: string
        description:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
      required: [name, command, args]
    UpdateMcpDefinitionRequest:
      type: object
      properties:
        name:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        description:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
    McpSetItem:
      type: object
      properties:
        serverId:
          type: string
          format: uuid
        disabled:
          type: boolean
      required: [serverId]
    McpSet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/McpSetItem'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, items, isActive, createdAt, updatedAt]
    CreateMcpSetRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/McpSetItem'
        description:
          type: string
      required: [name, items]
    UpdateMcpSetRequest:
      type: object
      properties:
        name:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/McpSetItem'
        description:
          type: string
...
    SaveConfigRequest:
      type: object
      properties:
        path:
          type: string
          minLength: 1
        content:
          type: string
      required: [path, content]
...